Separate test_models into single modules per model.
