+++ tests/test_flask/test_models/test_Meter.py
@@ -1,0 +1,9 @@
+"""Test the Meter model."""
+
+from run.models import Meter
+
+
+def test_Meter_repr():
+    """Test the __repr__ method."""
+    assert repr(
+        Meter(name='name', pattern='pattern')) == "<Meter 'name'>"
