+++ tests/test_flask/test_models/test_Poet.py
@@ -1,0 +1,10 @@
+"""Test the Poet model."""
+
+from run.models import Poet
+
+
+def test_Poet_repr():
+    """Test the __repr__ method."""
+    assert repr(
+        Poet(name='name')
+    ) == "<Poet 'name'>"
